Extract useSortFilter hook, remove duplicated sort/filter logic (closes #128)
Both AllContacts and ProjectView had identical sort state, filter state,
setFilter, and handleSort. The shared hook also exposes resetAll(),
replacing the scattered setSort/setFilters reset calls.
